+++ src/components/m-sticky/_example.html
@@ -1,0 +1,1 @@
+<!--Please create here a HTML example by using just default HTML tags-->
