+++ src/components/m-sticky/_preview.html
@@ -1,0 +1,1 @@
+<axa-sticky classes="m-sticky"></axa-sticky>
