+++ src/js/react-exports.js
@@ -158,3 +158,6 @@ export const AXATestimonials = withReact(AXATestimonialsWC);
 
 import AXATableWC from '../components/o-table/';
 export const AXATable = withReact(AXATableWC);
+
+import AXAStickyWC from '../components/m-sticky/';
+export const AXASticky = withReact(AXAStickyWC);
